+++ src/test/java/org/fugerit/java/client/redis/TestClientRedisRun.java
@@ -1,19 +1,31 @@
 package org.fugerit.java.client.redis;
 
 
+import com.github.stefanbirkner.systemlambda.SystemLambda;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.fop.cli.Main;
+import org.fugerit.java.core.cfg.ConfigRuntimeException;
 import org.fugerit.java.core.cli.ArgUtils;
+import org.fugerit.java.core.function.SafeFunction;
+import org.fugerit.java.core.lang.ex.CodeException;
+import org.fugerit.java.core.lang.ex.CodeRuntimeException;
+import org.fugerit.java.tool.util.MainHelper;
 import org.junit.Assert;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
+import java.security.Permission;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.function.IntConsumer;
 
 @Testcontainers
 @Slf4j
@@ -26,7 +38,27 @@ class TestClientRedisRun {
     @Container
     GenericContainer redis = new GenericContainer(DockerImageName.parse("redis:7.2-alpine")).withExposedPorts(6379);
 
+    private IntConsumer originalExitAction;
+
+    @BeforeEach
+    void preSetup() {
+        this.originalExitAction = MainHelper.getDefaultExitAction();
+        MainHelper.setDefaultExitAction( ec -> {
+            if ( ec == MainHelper.OK_DEFAULT ) {
+                log.info( "exit ok!" );
+            } else {
+                throw new CodeRuntimeException( ec );
+            }
+        } );
+    }
+
+    @AfterEach
+    void postSetup() {
+        MainHelper.setDefaultExitAction( this.originalExitAction );
+    }
+
     private String[] getParameters( String... additionalParams ) {
+
         String address = redis.getHost();
         Integer port = redis.getFirstMappedPort();
         String redisUrl = "redis://"+address+":"+port;
@@ -67,31 +99,25 @@ class TestClientRedisRun {
 
     @Test
     void testModeNotSupported() {
-        boolean ok = true;
         try {
             // modeUnkown
             String[] modeUnkown = getParameters( ArgUtils.getArgString( ClientRedisArgs.ARG_MODE ), "unknown" );
             ClientRedisRun.main( modeUnkown );
-        } catch (Exception e) {
-            log.warn( "Errore in unit test : "+e ,e );
-            ok = false;
+        } catch (CodeRuntimeException e) {
+            log.info( "Exit code : {}", e.getCode() );
+            Assertions.assertEquals( MainHelper.FAIL_MISSING_REQUIRED_PARAM, e.getCode() );
         }
-        Assert.assertTrue( ok );
     }
 
     @Test
     void testWrongConfiguration() {
-        boolean ok = true;
         try {
-            // modeUnkown
-            String[] wrongConfiguration = getParameters( ArgUtils.getArgString( ClientRedisArgs.ARG_MODE ), ClientRedisArgs.MODE_SINGLE_COMMAND );
-            ClientRedisRun.main( wrongConfiguration );
-        } catch (Exception e) {
-            log.warn( "Errore in unit test : "+e ,e );
-            ok = false;
+            String[] wrongConfiguration = getParameters(ArgUtils.getArgString(ClientRedisArgs.ARG_MODE), ClientRedisArgs.MODE_SINGLE_COMMAND);
+            ClientRedisRun.main(wrongConfiguration);
+        } catch ( CodeRuntimeException e ) {
+            log.info( "Exit code : {}", e.getCode() );
+            Assertions.assertEquals( MainHelper.FAIL_MISSING_REQUIRED_PARAM, e.getCode() );
         }
-        Assert.assertTrue( ok );
     }
-
 
 }